[feature] Timeline cursor
Added timeline cursor that can be dragged across the timeline. The cursor appears in front of the timeline and the keyframes. Changed the keyframe color to blue, so it is easier to see the red cursor.
Store field references to the timeline and cursor GraphicsItems in the KeyframeEditorScene. Add a update scene size and set timeline length option in the scene to adjust the respective GraphicsItems.