[change] Timeline related Keyframe and Cursor adjustments
The KeyframeItem instances store a reference to the timeline to appropriatly position themselves.
Keyframes also indicate their position on the timeline with a tic mark below their thumbnail.
They timeline cursor also holds a reference to the timeline to bound it's movement to the
timeline. Clicking on the timeline will move the cursor to that position on the timeline.