[feature] Keyframe ojbect in animation.py
The Keyframe object in the animation module is an object that stores reference to a scene,
thumbnail, and timestamp. The keyframe object is designed to be a specific stage of an
animation and these objects are stored chronologically in the animations manager. When
a new keyframe is created, a new scene with a matching name is created, the timestamp is
set, and a thumbnail from the current graphics view is saved.