[change] Keyframe implementation into animation scaffolding
The new Keyfram objects are added into the scaffolding and test material
from the rest of the project. The animations manager has an updated system
for holding a collection of Keyframes and sorts them. Keyframes inserted
into edit and delete commands. Timeline related checks for last keyframe
and inserting and deleting time use Keyframe objects. ChimeraX api standards
for animations reset state and take and restore snapshot is also updated for
Keyframe objects.