[change] Keyfram Graphics Ojbect timeline placements
Factor half width into calculating the time info label on the Keyframe Graphic Opbject.
Fix check to see if keyframe exists why trying to edit in Animation manager.
Implement a default size class const in KeyframeItem and use it to properly calculate the x position when placing a new keyframe on the timeline.