[feature] Trigger for keyframe add interaction with gui
Move adding a keyframe graphics object for the scene into it's own method. This will make it easier to have a listener to add new keyframes in the future.
Getter for triggers from the animation manager.
Pass a reference to the triggerset from the animation manager into the keyframe editor widget.
Keyframes appear in the gui tool when they are created in the command line. Added handler for keyframe added trigger in KeyframeEditor widget.