[change] Animation length setting
Rename the set length for the timeline graphics item group to set time length. Also fix improper round when setting the pix_length in that method.
Use setter for adjusting the length of an animation in the animation manager. This means we can call the length change trigger in only one spot.
Timeline length changes are updated with handlers in the gui when the length is changed in the animations manager.