[feature] Animations manager fires preview trigger
Add preview trigger to animations manager for when a preview at a certain time is called for.
Fix bug in previewing where time could round to try and show a frame that was outside the len bound of the calculated frames. Fixed by if the frame rounded up ever that it picks the last frame to show anyways.