Update in CallForNFrames done() method.
Previously calling done() on a CallForNFrames more than once caused an error because it would try and remove itself from the session motion_in_progress attr when it did not exist there. This commit adds a simple check to make sure that it does not attempt a remove if it has already been removed and avoids an error when you call .done() on a CallForNFrames more than once. This fix is relevant for the animations bundle which needs to be able to call .done() when trying to stop playing queued frames for pause. The pause command doesn't know if .done() has or has not been called automatically by CallForNFrames, so it needs this safety to avoid potential errors.