[feature] Setup for SceneVisibility manager
Initial setup for the SceneVisiblity state manager. Will be used to
manage data on visitiblity state of objects in the scene.

Set up scaffolding for a SceneVisibility state manager.

Set up scene visibility in the scene manager and main scene class.

Add model visibility to SceneVisibility. Interpolate and interpolatable is not implemented yet.

Implement remove models to SceneVisibility.

TODO marker for next implementation for SceneVisibility.